+++ src/main/java/com/codefactorygroup/betting/service/implementations/MarketServiceImpl.java
@@ -1,38 +1,36 @@
 package com.codefactorygroup.betting.service.implementations;
 
 import com.codefactorygroup.betting.converter.MarketDTOtoMarketConverter;
-import com.codefactorygroup.betting.converter.SelectionDTOtoSelectionConverter;
+import com.codefactorygroup.betting.domain.Event;
 import com.codefactorygroup.betting.domain.Market;
 import com.codefactorygroup.betting.dto.MarketDTO;
-import com.codefactorygroup.betting.dto.SelectionDTO;
 import com.codefactorygroup.betting.exception.EntityAlreadyExistsException;
 import com.codefactorygroup.betting.exception.NoSuchEntityExistsException;
+import com.codefactorygroup.betting.repository.EventRepository;
 import com.codefactorygroup.betting.repository.MarketRepository;
 import com.codefactorygroup.betting.service.MarketService;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
-import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service(value = "marketService")
 public class MarketServiceImpl implements MarketService {
 
     private final MarketRepository marketRepository;
 
+    private final EventRepository eventRepository;
     private final MarketDTOtoMarketConverter marketDTOtoMarketConverter;
 
-    private final SelectionDTOtoSelectionConverter selectionDTOtoSelectionConverter;
-
-
-    public MarketServiceImpl(MarketRepository marketRepository, MarketDTOtoMarketConverter marketDTOtoMarketConverter, SelectionDTOtoSelectionConverter selectionDTOtoSelectionConverter) {
+    public MarketServiceImpl(MarketRepository marketRepository, EventRepository eventRepository,
+                             MarketDTOtoMarketConverter marketDTOtoMarketConverter) {
         this.marketRepository = marketRepository;
+        this.eventRepository = eventRepository;
         this.marketDTOtoMarketConverter = marketDTOtoMarketConverter;
-        this.selectionDTOtoSelectionConverter = selectionDTOtoSelectionConverter;
     }
 
-    @Transactional
     @Override
     public MarketDTO getMarket(Integer marketId) {
         return marketRepository.findById(marketId)
@@ -40,31 +38,58 @@ public class MarketServiceImpl implements MarketService {
                 .orElseThrow(() -> new NoSuchEntityExistsException(String.format("Market with ID=%d doesn't exist.", marketId)));
     }
 
+    @Override
+    public List<MarketDTO> getAllMarkets() {
+        return marketRepository.findAll()
+                .stream()
+                .map(MarketDTO::converter)
+                .collect(Collectors.toList());
+    }
+
+    @Override
+    public List<MarketDTO> getMarketsByEventId(Integer eventId) {
+        Optional<Event> eventOptional = eventRepository.findById(eventId);
+        if (eventOptional.isEmpty()) {
+            throw new NoSuchEntityExistsException(String.format("Event with ID=%d doesn't exist.", eventId));
+        }
+        return marketRepository.findMarketsByEventId(eventId)
+                .stream()
+                .map(MarketDTO::converter)
+                .collect(Collectors.toList());
+    }
+
     @Transactional
     @Override
-    public MarketDTO addMarket(MarketDTO market) {
-        boolean foundMarket = marketRepository.existsById(market.id());
-        if (foundMarket) {
-            throw new EntityAlreadyExistsException(String.format("Market with ID=%d already exists.", market.id()));
+    public MarketDTO addMarket(final Integer eventId, final MarketDTO newMarket) {
+        Event event = eventRepository.findById(eventId)
+                .orElseThrow(() -> new NoSuchEntityExistsException(String.format("Event with ID=%d doesn't exist.", eventId)));
+
+        boolean existsMarket = marketRepository.existsByEventIdAndName(eventId, newMarket.name());
+        if (existsMarket) {
+            throw new EntityAlreadyExistsException(String.format("Market with name=%s already exists.", newMarket.name()));
+        } else {
+            Market market = marketDTOtoMarketConverter.convert(newMarket);
+            event.addMarket(market);
+
+            return MarketDTO.converter(
+                    marketRepository.save(market));
         }
-        return MarketDTO.converter(marketRepository.save(marketDTOtoMarketConverter.convert(market)));
     }
+
 
     @Transactional
     @Override
     public void deleteMarket(Integer marketId) {
-        marketRepository.deleteById(marketId);
+        boolean marketExists = marketRepository.existsById(marketId);
+        if (marketExists) {
+            marketRepository.deleteById(marketId);
+        } else {
+            throw new NoSuchEntityExistsException(String.format("Market with ID=%d doesn't exist.", marketId));
+        }
     }
 
     private Market update(final Market market, final MarketDTO toUpdateMarket) {
-        List<SelectionDTO> selectionDTOS = Optional.of(toUpdateMarket).map(MarketDTO::selections).orElseGet(Collections::emptyList);
-        market.setId(toUpdateMarket.id());
         market.setName(toUpdateMarket.name());
-        market.setSelections(selectionDTOS
-                .stream()
-                .map(selectionDTOtoSelectionConverter::convert)
-                .toList()
-        );
         return market;
     }
