+++ src/main/java/com/codefactorygroup/betting/service/implementations/SportServiceImpl.java
@@ -1,9 +1,7 @@
 package com.codefactorygroup.betting.service.implementations;
 
-import com.codefactorygroup.betting.converter.CompetitionDTOtoCompetitionConverter;
 import com.codefactorygroup.betting.converter.SportDTOtoSportConverter;
 import com.codefactorygroup.betting.domain.Sport;
-import com.codefactorygroup.betting.dto.CompetitionDTO;
 import com.codefactorygroup.betting.dto.SportDTO;
 import com.codefactorygroup.betting.exception.EntityAlreadyExistsException;
 import com.codefactorygroup.betting.exception.NoSuchEntityExistsException;
@@ -12,9 +10,7 @@ import com.codefactorygroup.betting.service.SportService;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
-import java.util.Collections;
 import java.util.List;
-import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service(value = "sportService")
@@ -24,48 +20,53 @@ public class SportServiceImpl implements SportService {
 
     private final SportDTOtoSportConverter sportDTOtoSportConverter;
 
-    private final CompetitionDTOtoCompetitionConverter competitionDTOtoCompetitionConverter;
 
-    public SportServiceImpl(SportRepository sportRepository, SportDTOtoSportConverter sportDTOtoSportConverter, CompetitionDTOtoCompetitionConverter competitionDTOtoCompetitionConverter) {
+    public SportServiceImpl(SportRepository sportRepository, SportDTOtoSportConverter sportDTOtoSportConverter) {
         this.sportRepository = sportRepository;
         this.sportDTOtoSportConverter = sportDTOtoSportConverter;
-        this.competitionDTOtoCompetitionConverter = competitionDTOtoCompetitionConverter;
     }
 
-    @Transactional
     @Override
     public SportDTO getSport(Integer sportId) {
         return sportRepository.findById(sportId)
                 .map(SportDTO::converter)
-                .orElseThrow(() -> new NoSuchEntityExistsException(String.format("Sport with ID=%d doesn't exists.", sportId)));
+                .orElseThrow(() -> new NoSuchEntityExistsException(String.format("Sport with ID=%d doesn't exist.", sportId)));
     }
 
+    @Override
+    public List<SportDTO> getAllSports() {
+        return sportRepository.findAll()
+                .stream()
+                .map(SportDTO::converter)
+                .collect(Collectors.toList());
+    }
 
     @Transactional
     @Override
-    public SportDTO addSport(SportDTO sport) {
-        boolean foundSport = sportRepository.existsById(sport.id());
-        if (foundSport) {
-            throw new EntityAlreadyExistsException(String.format("Sport with ID=%d already exists.", sport.id()));
+    public SportDTO addSport(SportDTO newSport) {
+        boolean existsSport = sportRepository.existsByName(newSport.name());
+        if (existsSport) {
+            throw new EntityAlreadyExistsException(String.format("Sport with name=%s already exists.", newSport.name()));
+        } else {
+            Sport sport = sportDTOtoSportConverter.convert(newSport);
+            return SportDTO.converter(
+                    sportRepository.save(sport));
         }
-        return SportDTO.converter(sportRepository.save(sportDTOtoSportConverter.convert(sport)));
     }
-
 
     @Transactional
     @Override
     public void deleteSport(Integer sportId) {
-        sportRepository.deleteById(sportId);
+        boolean sportExists = sportRepository.existsById(sportId);
+        if (sportExists) {
+            sportRepository.deleteById(sportId);
+        } else {
+            throw new NoSuchEntityExistsException(String.format("Sport with ID=%d doesn't exist.", sportId));
+        }
     }
 
     private Sport update(final Sport sport, final SportDTO toUpdateSport) {
-        List<CompetitionDTO> competitionDTOList = Optional.of(toUpdateSport).map(SportDTO::competitions).orElseGet(Collections::emptyList);
-        sport.setId(toUpdateSport.id());
         sport.setName(toUpdateSport.name());
-        sport.setCompetitions(competitionDTOList
-                .stream()
-                .map(competitionDTOtoCompetitionConverter::convert)
-                .collect(Collectors.toList()));
         return sport;
     }
 
@@ -79,4 +80,5 @@ public class SportServiceImpl implements SportService {
                 .map(SportDTO::converter)
                 .orElseThrow(() -> new NoSuchEntityExistsException(String.format("Sport with ID=%d doesn't exist.", sportId)));
     }
+
 }
